+++ src/earthkit/utils/decorators/__init__.py
@@ -8,7 +8,7 @@
 
 """Function decorators and wrappers for use in the downstream EarthKit packages."""
 
-from earthkit.utils.decorators.dispatch import dispatch
+from earthkit.utils.decorators._dispatch import dispatch
 from earthkit.utils.decorators.experimental import ExperimentalWarning, experimental
 from earthkit.utils.decorators.format_handlers import format_handler
 from earthkit.utils.decorators.thread_handlers import thread_safe_cached_property
