+++ tests/test_dispatch.py
@@ -14,7 +14,8 @@ import pytest
 import xarray as xr
 from earthkit.data import SimpleFieldList
 
-from earthkit.utils.decorators.dispatch import (
+from earthkit.utils.decorators import _dispatch as dispatch_module
+from earthkit.utils.decorators._dispatch import (
     ArrayDispatcher,
     ArrayLikeDispatcher,
     FieldListDispatcher,
@@ -133,7 +134,7 @@ class TestXArrayDispatcher:
         mock_module = MagicMock()
         mock_module.test_func = MagicMock(return_value="xarray_result")
 
-        with patch("earthkit.utils.decorators.dispatch.import_module", return_value=mock_module):
+        with patch.object(dispatch_module, "import_module", return_value=mock_module):
             result = dispatcher.dispatch("test_func", "dummy.module", 1, 2, key="value")
 
         assert result == "xarray_result"
@@ -163,7 +164,7 @@ class TestFieldListDispatcher:
         mock_module = MagicMock()
         mock_module.test_func = MagicMock(return_value="fieldlist_result")
 
-        with patch("earthkit.utils.decorators.dispatch.import_module", return_value=mock_module):
+        with patch.object(dispatch_module, "import_module", return_value=mock_module):
             result = dispatcher.dispatch("test_func", "dummy.module", 1, 2, key="value")
 
         assert result == "fieldlist_result"
@@ -251,7 +252,7 @@ class TestArrayLikeDispatcher:
         mock_module = MagicMock()
         mock_module.test_func = MagicMock(return_value="array_result")
 
-        with patch("earthkit.utils.decorators.dispatch.import_module", return_value=mock_module) as mock_import:
+        with patch.object(dispatch_module, "import_module", return_value=mock_module) as mock_import:
             result = dispatcher.dispatch("test_func", "dummy.module", [1, 2, 3])
 
         mock_import.assert_called_once_with("dummy.module.array")
@@ -271,7 +272,7 @@ class TestDispatchWrapperArrayLike:
         mock_module = MagicMock()
         mock_module.process_data = MagicMock(return_value="array_result")
 
-        with patch("earthkit.utils.decorators.dispatch.import_module", return_value=mock_module):
+        with patch.object(dispatch_module, "import_module", return_value=mock_module):
             result = process_data([1, 2, 3])
 
         assert result == "array_result"
@@ -286,7 +287,7 @@ class TestDispatchWrapperArrayLike:
         mock_module = MagicMock()
         mock_module.process_data = MagicMock(return_value="array_result")
 
-        with patch("earthkit.utils.decorators.dispatch.import_module", return_value=mock_module):
+        with patch.object(dispatch_module, "import_module", return_value=mock_module):
             result = process_data(42)
 
         assert result == "array_result"
@@ -301,7 +302,7 @@ class TestDispatchWrapperArrayLike:
         mock_module = MagicMock()
         mock_module.process_data = MagicMock(return_value="array_result")
 
-        with patch("earthkit.utils.decorators.dispatch.import_module", return_value=mock_module):
+        with patch.object(dispatch_module, "import_module", return_value=mock_module):
             result = process_data(3.14)
 
         assert result == "array_result"
@@ -338,7 +339,7 @@ class TestArrayDispatcher:
         mock_module = MagicMock()
         mock_module.test_func = MagicMock(return_value="array_result")
 
-        with patch("earthkit.utils.decorators.dispatch.import_module", return_value=mock_module):
+        with patch.object(dispatch_module, "import_module", return_value=mock_module):
             result = dispatcher.dispatch("test_func", "dummy.module", 1, 2, key="value")
 
         assert result == "array_result"
@@ -359,7 +360,7 @@ class TestDispatchWrapper:
         mock_xarray_module = MagicMock()
         mock_xarray_module.process_data = MagicMock(return_value="xarray_implementation")
 
-        with patch("earthkit.utils.decorators.dispatch.import_module", return_value=mock_xarray_module):
+        with patch.object(dispatch_module, "import_module", return_value=mock_xarray_module):
             result = process_data(TEST_XARRAY_DATAARRAY)
 
         assert result == "xarray_implementation"
@@ -375,7 +376,7 @@ class TestDispatchWrapper:
         mock_array_module = MagicMock()
         mock_array_module.process_data = MagicMock(return_value="array_implementation")
 
-        with patch("earthkit.utils.decorators.dispatch.import_module", return_value=mock_array_module):
+        with patch.object(dispatch_module, "import_module", return_value=mock_array_module):
             result = process_data(TEST_NUMPY_ARRAY)
 
         assert result == "array_implementation"
@@ -390,7 +391,7 @@ class TestDispatchWrapper:
         mock_array_module = MagicMock()
         mock_array_module.process_data = MagicMock(return_value="array_implementation")
 
-        with patch("earthkit.utils.decorators.dispatch.import_module", return_value=mock_array_module):
+        with patch.object(dispatch_module, "import_module", return_value=mock_array_module):
             result = process_data(TEST_NUMPY_ARRAY)
             assert result == "array_implementation"
 
@@ -404,7 +405,7 @@ class TestDispatchWrapper:
         mock_xarray_module = MagicMock()
         mock_xarray_module.process_data = MagicMock(return_value="xarray_implementation")
 
-        with patch("earthkit.utils.decorators.dispatch.import_module", return_value=mock_xarray_module):
+        with patch.object(dispatch_module, "import_module", return_value=mock_xarray_module):
             result = process_data(TEST_XARRAY_DATAARRAY, other_param=42)
 
         assert result == "xarray_implementation"
@@ -420,7 +421,7 @@ class TestDispatchWrapper:
         mock_xarray_module = MagicMock()
         mock_xarray_module.process_with_name_match = MagicMock(return_value="xarray_implementation")
 
-        with patch("earthkit.utils.decorators.dispatch.import_module", return_value=mock_xarray_module):
+        with patch.object(dispatch_module, "import_module", return_value=mock_xarray_module):
             result = process_with_name_match(TEST_XARRAY_DATAARRAY)
 
         assert result == "xarray_implementation"
@@ -472,7 +473,7 @@ class TestDispatchWrapper:
         mock_xarray_module = MagicMock()
         mock_xarray_module.process_data = MagicMock(return_value="xarray_with_kwargs")
 
-        with patch("earthkit.utils.decorators.dispatch.import_module", return_value=mock_xarray_module):
+        with patch.object(dispatch_module, "import_module", return_value=mock_xarray_module):
             result = process_data(TEST_XARRAY_DATAARRAY, multiplier=3)
 
         assert result == "xarray_with_kwargs"
@@ -510,7 +511,7 @@ class TestDispatchWrapper:
         mock_array_module = MagicMock()
         mock_array_module.with_only_array = MagicMock(return_value="array_dispatched")
 
-        with patch("earthkit.utils.decorators.dispatch.import_module", return_value=mock_array_module):
+        with patch.object(dispatch_module, "import_module", return_value=mock_array_module):
             result = with_only_array(TEST_NUMPY_ARRAY)
             assert result == "array_dispatched"
 
@@ -518,21 +519,21 @@ class TestDispatchWrapper:
         mock_xarray_module = MagicMock()
         mock_xarray_module.with_all_defaults = MagicMock(return_value="xarray_dispatched")
 
-        with patch("earthkit.utils.decorators.dispatch.import_module", return_value=mock_xarray_module):
+        with patch.object(dispatch_module, "import_module", return_value=mock_xarray_module):
             result = with_all_defaults(TEST_XARRAY_DATAARRAY)
 
         assert result == "xarray_dispatched"
         mock_xarray_module.with_all_defaults.assert_called_once_with(TEST_XARRAY_DATAARRAY)
 
         # xarray input is array-like, so should be dispatched
-        with patch("earthkit.utils.decorators.dispatch.import_module", return_value=mock_xarray_module):
+        with patch.object(dispatch_module, "import_module", return_value=mock_xarray_module):
             with_only_array_and_array_like(TEST_XARRAY_DATAARRAY)
 
         assert result == "xarray_dispatched"
         mock_xarray_module.with_all_defaults.assert_called_once_with(TEST_XARRAY_DATAARRAY)
 
         # With xarray is not a strict array, so should not be dispatched
-        with patch("earthkit.utils.decorators.dispatch.import_module", return_value=mock_xarray_module):
+        with patch.object(dispatch_module, "import_module", return_value=mock_xarray_module):
             with pytest.raises(TypeError, match="No dispatcher matched for function"):
                 with_only_array(TEST_XARRAY_DATAARRAY)
 
@@ -551,7 +552,7 @@ class TestDispatchIntegration:
         mock_xarray_module = MagicMock()
         mock_xarray_module.process = MagicMock(return_value="xarray")
 
-        with patch("earthkit.utils.decorators.dispatch.import_module", return_value=mock_xarray_module):
+        with patch.object(dispatch_module, "import_module", return_value=mock_xarray_module):
             result = process(TEST_XARRAY_DATAARRAY)
 
         assert result == "xarray"
@@ -582,7 +583,7 @@ class TestDispatchIntegration:
         mock_xarray_module = MagicMock()
         mock_xarray_module.process_with_defaults = MagicMock(return_value="xarray")
 
-        with patch("earthkit.utils.decorators.dispatch.import_module", return_value=mock_xarray_module):
+        with patch.object(dispatch_module, "import_module", return_value=mock_xarray_module):
             # Call with only required argument
             result = process_with_defaults(TEST_XARRAY_DATAARRAY)
             assert result == "xarray"
